Add onSelectedVersion callback so it can be handled in any way the caller wishes
